Optimización upload fotos

+++ app/src/main/java/com/dataservicios/ttauditalicorpcp/services/UpdateService.java
@@ -6,31 +6,50 @@ import android.content.Intent;
 import android.os.IBinder;
 import android.support.annotation.Nullable;
 import android.util.Log;
+import android.widget.Toast;
 
+import com.androidnetworking.AndroidNetworking;
+import com.androidnetworking.common.ANRequest;
+import com.androidnetworking.common.ANResponse;
+import com.androidnetworking.common.Priority;
+import com.androidnetworking.error.ANError;
+import com.androidnetworking.interfaces.JSONObjectRequestListener;
+import com.androidnetworking.interfaces.UploadProgressListener;
+import com.dataservicios.ttauditalicorpcp.R;
 import com.dataservicios.ttauditalicorpcp.app.AppController;
 import com.dataservicios.ttauditalicorpcp.model.Media;
 import com.dataservicios.ttauditalicorpcp.repo.MediaRepo;
 import com.dataservicios.ttauditalicorpcp.util.AuditUtil;
+import com.dataservicios.ttauditalicorpcp.util.BitmapLoader;
 import com.dataservicios.ttauditalicorpcp.util.Connectivity;
+import com.dataservicios.ttauditalicorpcp.util.GlobalConstant;
+
+
+import org.json.JSONException;
+import org.json.JSONObject;
+
+import java.io.File;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.UUID;
+
+import okhttp3.Response;
 
 public class UpdateService extends Service {
 
-
-    private final String TAG = UpdateService.class.getSimpleName();
+    private final String LOG_TAG = UpdateService.class.getSimpleName();
     private final Integer contador = 0;
-
     private Context context = this;
-
-    static final int DELAY =  60000; //2 minutos de espera
-    //static final int DELAY = 9000; //9 segundo de espera
+    static final int DELAY =  60000; //1 minutos de espera
+//    static final int DELAY = 9000; // segundo de espera
     private boolean runFlag = false;
     private Updater updater;
-
     private AppController application;
-
     private MediaRepo mediaRepo;
     private Media media;
+    private ArrayList<Media> medias;
     private AuditUtil auditUtil;
+    private File file;
 
     public UpdateService() {
     }
@@ -48,7 +67,7 @@ public class UpdateService extends Service {
         mediaRepo = new MediaRepo(this);
         media = new Media();
         auditUtil = new AuditUtil(context);
-        Log.d(TAG, "onCreated");
+                Log.d(LOG_TAG, "onCreated");
     }
 
     @Override
@@ -59,7 +78,7 @@ public class UpdateService extends Service {
         application.setServiceRunningFlag(false);
         updater.interrupt();
         updater = null;
-        Log.d(TAG, "onDestroyed");
+                Log.d(LOG_TAG, "onDestroyed");
     }
 
     @Override
@@ -70,7 +89,7 @@ public class UpdateService extends Service {
             updater.start();
         }
 
-        Log.d(TAG, "onStarted");
+                Log.d(LOG_TAG, "onStarted");
         return START_STICKY;
     }
 
@@ -79,45 +98,64 @@ public class UpdateService extends Service {
             super("UpdaterService-UpdaterThread");
         }
 
-
         @Override
         public void run() {
 
             UpdateService updaterService = UpdateService.this;
             while (updaterService.runFlag) {
-                Log.d(TAG, "UpdaterThread running");
+                Log.d(LOG_TAG, "UpdaterThread running");
                 try{
-
                     if(Connectivity.isConnected(context)) {
                         if (Connectivity.isConnectedFast(context)) {
-
-                            Log.i(TAG," Conexión rápida" );
+                                    Log.i(LOG_TAG," Conexión rápida" );
                             if (mediaRepo.countReg() >0 ) {
-                                media = (Media) mediaRepo.findFirstReg();
-                                boolean response = auditUtil.uploadMedia(media,1);
-                                if (response) {
-                                    mediaRepo.delete(media);
-                                    Log.i(TAG," Send success images database server and delete local database and file " );
+                                media.setId(0);
+                                medias = (ArrayList<Media>) mediaRepo.findAll();
+                                for (Media m: medias){
+                                    file = null;
+                                    file = new File(BitmapLoader.getAlbumDirTemp(context).getAbsolutePath() + "/" + m.getFile());
+                                    if(file.exists()){
+                                        media = m;
+                                        break;
+                                    }
+                                }
+//
+                                if (media.getId() != 0){
+                                    // NOTA eliminar  de "auditUtil.uploadMedia"
+                                    // la eliminación de archivos
+                                    //  file.delete() para controlar la eliminación en base de datos
+                                    //boolean response = auditUtil.uploadMedia(media,1);
+                                    boolean response = auditUtil.sendUploadPhotoServer(media);
+                                    if (response) {
+                                        file = null;
+                                        file = new File(BitmapLoader.getAlbumDirTemp(context).getAbsolutePath() + "/" + media.getFile());
+                                        if(file.exists()){
+                                            file.delete();
+                                            mediaRepo.delete(media);
+                                        }
+                                        Log.i(LOG_TAG," Se envió correctamente los datos al servidor, se eliminó el registro en la base de datos, y se eliminó  el archivo " );
+                                    } else {
+                                        Log.i(LOG_TAG," El servidor responde falso, no se pudo enviar el archivo al servidor " );
+                                    }
                                 }
                             } else{
-                                Log.i(TAG, "No found records in media table for send");
+                                        Log.i(LOG_TAG, "No se encontró registros media para el envío");
                             }
 
                         }else {
-                            Log.i(TAG," Connectivity slow" );
+                                    Log.i(LOG_TAG," COnexión a internet Lenta" );
                         }
                     } else {
-                        Log.i(TAG," No internet connection" );
+                                Log.i(LOG_TAG,"No hay conexión a internet" );
                     }
                     Thread.sleep(DELAY);
                 }catch(InterruptedException e){
                     updaterService.runFlag = false;
                     application.setServiceRunningFlag(true);
                 }
-
             }
         }
-
-
     }
 }
+
+
